reorder method execution
reason:

until now, we were 1) updating db and 2) calling pushToIxsi
asyncly. the problem is that pushToIxsi logic relies on db for the old
state and dto for the new state. it compares the two whether a push
should be executed or not. but since we first were updating db, the old
state was lost and never available to pushToIxsi.

with this change, we fix this by first executing ixsi logic/push and
then db update. therefore, we must also make sure that ixsi errors do
not affect db update (which is more important) and catch all exceptions.