extend "sequential operation" idea to all messages coming from a station
reason:

it's not only StationStatusDTO that we care about. intertwined operation
dependencies like a status update which sets the pedelec to inoperative
and a getAvailablePedelecs should not be executed in parallel, because
status update affects available pedelecs.

we hereby migrate from a producer/consumer pattern to a locking/sync
mechanism. making producer/consumer work for such a requirement would
introduce complexity. the locks are not global, but per station, so we
only serialize the operations per station.